Les clics gauche en jeu tirent des projectiles vers le curseur
rcz.panel.GamePanel: Appel de la fonction draw dans ProjectileStock
control.CameControler: modifs a ignorer pour le moment
control: ajout MouseControler
rcz.Window: prise en compte du MouseListener
entity.creature.Player: ajout fonction Attack()
entity: Ajout Projectile
entity.Entity: x et y sont des types double
entity.creature.Creature: suppression de dX et dY
entity.creature.Player: ajustements en consequence

+++ src/entity/Entity.java
@@ -4,7 +4,7 @@ import java.awt.Graphics;
 
 public abstract class Entity {
 
-	protected int x, y;
+	protected double x, y;
 	protected int distanceToPlayer;
 	
 	protected Entity(int posX, int posY) {
